Harden curated file scope
